requirements.txt
keyboard==0.13.5
lxml==5.3.1
MouseInfo==0.1.3
mss==10.0.0
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-docx==1.1.2
pytweening==1.2.0
typing_extensions==4.12.2
